+++ packages/react-components/src/components/CheckboxGroup/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./CheckboxGroup";
+export type { CheckboxGroupProps } from "./CheckboxGroup";