+++ packages/react-components/src/components/Form/Form.tsx
@@ -3,8 +3,6 @@ import {
   FormProps as ReactAriaFormProps,
 } from "react-aria-components";
 
-export type FormProps = ReactAriaFormProps;
-
-export default function Form({ ...props }: FormProps) {
+export default function Form({ ...props }: ReactAriaFormProps) {
   return <ReactAriaForm {...props} />;
 }