+++ packages/react-components/src/components/RadioGroup/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./RadioGroup";
+export type { RadioGroupProps } from "./RadioGroup";