+++ packages/react-components/src/pages/RadioGroup/index.ts
@@ -1,0 +1,3 @@
+import RadioGroupPage from "./RadioGroup";
+
+export default RadioGroupPage;