+++ packages/react-components/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 0.2.3
+
+### Changed
+
+- Switch component gets `prefers-reduced-motion` handling to turn off animation for users that prefer reduced motion.
+- Switch component `cursor` styling applies to `children` text in addition to visual switch component.
+
+###
+
+- Added RadioGroup/Radio components.
+
+## 0.2.2
+
+### Added
+
+- Added CheckboxGroup/Checkbox components.
+
 ## 0.2.1
 
 No changes from v0.2.0.